+++ src/test/java/ulaval/glo2003/application/product/ProductFactoryTest.java
@@ -16,8 +16,8 @@ public class ProductFactoryTest {
     private static final String SELLER_ID = "1";
     private static final String TITLE = "titleOfProduct";
     private static final String DESCRIPTION = "descriptionOfProduct";
-    private static final Float SUGGESTED_PRICE = 7.7f;
-    private static final Float INVALID_SUGGESTED_PRICE = 0.5f;
+    private static final Double SUGGESTED_PRICE = 7.7;
+    private static final Double INVALID_SUGGESTED_PRICE = 0.5;
     private final List<String> CATEGORIES = new ArrayList<>() {
         {
             add("beauty");
